Fix findbugs/sonar violations in sfc-ios-xe-renderer
- Suspicious reference comparison of Boolean values
- Dead store to local variable
- Either log or rethrow this exception

Also enabled findbugs enforcement.

+++ sfc-renderers/sfc-ios-xe-renderer/src/main/java/org/opendaylight/sfc/iosxe/provider/renderer/IosXeRspProcessor.java
@@ -106,7 +106,7 @@ public class IosXeRspProcessor {
             return;
         }
         ServiceTypeChoice serviceTypeChoice = buildServiceFunctionChoice(serviceFunction);
-        Short serviceIndex = renderedServicePath.getStartingIndex();
+        short serviceIndex = renderedServicePath.getStartingIndex().shortValue();
         Services serviceEntry = createServicesEntry(serviceIndex, serviceTypeChoice);
         services.add(serviceEntry);
         serviceIndex--;
